+++ src/main/java/com/mvi/CSCB634College/major/MajorRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface MajorRepository extends JpaRepository<Major, Integer> {
+public interface MajorRepository extends JpaRepository<Major, Long> {
 
     List<Major> findAllByDepartment(Department department);
 }
